On master:

```Julia
  102.216 ms (5104687 allocations: 77.89 MiB)
  287.190 ms (5104702 allocations: 116.84 MiB)
julia> 
```

With this PR:
```Julia
 97.506 ms (5104687 allocations: 77.89 MiB)
  72.398 ms (5104706 allocations: 145.00 MiB)
```
So roughly ~4x faster decode, and 1.05x faster encode :) 

Benchmark code:
```Julia
using Random, BenchmarkTools, MsgPack
data = Dict(
	randstring(50) => rand(10^6),
	randstring(10) => rand(Float32, 2028 * 2024),
  42 => Dict("test" => 1f0)
);

test(data) = open("test.msgpack", "w") do io
	MsgPack.pack(io, data)
end
test() = open("test.msgpack") do io
	MsgPack.unpack(io)
end
@btime test($data)
@btime test()
```